Cycle fixes: extract `FocusContext` from comments
