+++ src/ui/components/Comments/TranscriptComments/types.ts
@@ -1,3 +1,4 @@
+import React from "react";
 import { Comment, Reply } from "ui/state/comments";
 
 export type CommentData =
@@ -5,3 +6,10 @@ export type CommentData =
   | { type: "new_comment"; comment: Comment }
   | { type: "reply"; comment: Reply }
   | { type: "new_reply"; comment: Reply };
+
+export const FocusContext = React.createContext({
+  autofocus: false,
+  isFocused: false,
+  blur: () => {},
+  close: () => {},
+});
